pyproject.toml
[tool.poetry]
name = "ai-examples"
version = "0.1.0"
description = ""
authors = ["Steve Gibbard <88480716+gibbardsteve@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "ai_examples"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.9"
langchain = "^0.0.348"
streamlit = "^1.27.2"
python-dotenv = "^1.0.0"
duckduckgo-search = "^3.9.11"
numexpr = "^2.8.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
flake8 = "^6.1.0"
black = "^23.7.0"
bandit = "^1.7.5"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
